+++ packages/expo-intent-launcher/src/IntentLauncher.ts
@@ -185,9 +185,24 @@ export async function startActivityAsync(
  * Opens an application by its package name.
  * @param packageName e.g., `com.google.android.gm` for Gmail.
  */
-export function openPackage(packageName: string) {
-  if (!ExpoIntentLauncher.openPackage) {
-    throw new UnavailabilityError('IntentLauncher', 'openPackage');
+export function openApplication(packageName: string) {
+  if (!ExpoIntentLauncher.openApplication) {
+    throw new UnavailabilityError('IntentLauncher', 'openApplication');
   }
-  return ExpoIntentLauncher.openPackage(packageName);
+  return ExpoIntentLauncher.openApplication(packageName);
+}
+
+/**
+ * Returns the icon of the specified application as a base64-encoded PNG image string.
+ * The returned string is prefixed with `data:image/png;base64,` and can be used directly in an `expo-image` `<Image>`
+ * component's `source` prop.
+ *
+ * @param packageName The package name of the target application, e.g. `com.google.android.gm` for Gmail.
+ * @return A promise that resolves to the base64-encoded PNG icon of the specified application, or an empty string if the icon could not be retrieved.
+ */
+export async function getApplicationIconAsync(packageName: string): Promise<string> {
+  if (!ExpoIntentLauncher.getApplicationIcon) {
+    throw new UnavailabilityError('IntentLauncher', 'getApplicationIconAsync');
+  }
+  return ExpoIntentLauncher.getApplicationIcon(packageName);
 }