+++ packages/expo-mail-composer/ios/MailComposerModule.swift
@@ -41,22 +41,22 @@ public class MailComposerModule: Module {
 
     // Function to check the availability of listed mail clients on the device
     Function("getClients") { () -> [MailClient] in
-      return self.getAvailableMailClients()
+      return getAvailableMailClients()
     }
   }
+}
 
-  // Private function to check each mail client for availability
-  private func getAvailableMailClients() -> [MailClient] {
-    var availableClients = [MailClient]()
-    for client in mailClients {
-      if let url = URL(string: client.url), UIApplication.shared.canOpenURL(url) {
-        availableClients.append(client)
-      }
+// Private function to check each mail client for availability
+private func getAvailableMailClients() -> [MailClient] {
+  var availableClients = [MailClient]()
+  for client in mailClients {
+    if let url = URL(string: client.url), UIApplication.shared.canOpenURL(url) {
+      availableClients.append(client)
     }
-    return availableClients
   }
+  return availableClients
+}
 
-  private func canSendMail() -> Bool {
-    return MFMailComposeViewController.canSendMail()
-  }
+private func canSendMail() -> Bool {
+  return MFMailComposeViewController.canSendMail()
 }